refactor domain product trust route
